requirements.txt
streamlit
pandas
fpdf
requests
python-dotenv
plotly
